+++ src/main/service/UserService.js
@@ -47,7 +47,14 @@ module.exports = {
     changeAuthConnection: async (connection, request) => {
         const { authcode } = request
         const userEmail = await UserDao.selectEmailByAuthCode(connection, authcode)
-        await UserDao.selectPermissionByEmail(connection, userEmail) //check permission > 변경하려는 사용자가 정식회원인지
+        if (userEmail.affectedRows == 0) {
+            throw Error("존재하지 않는 사용자입니다")
+        }
+        const checkPermission = await UserDao.selectPermissionByEmail(connection, userEmail) //check permission > 변경하려는 사용자가 정식회원인지
+
+        if (checkPermission[0].permission === 1) {
+            throw Error("이미 정식회원인 사용자입니다")
+        }
         const result = await UserDao.updateAuthcodeByEmail(connection, userEmail)
 
         if (!result || result.length < 1) {
@@ -64,25 +71,35 @@ module.exports = {
 
         const userResult = await UserDao.findUserByEmail(connection, email, password)
         if (!userResult || userResult.length < 1) {
-            throw Error("사용자 정보가 없습니다")
+            throw Error("존재하지 않는 사용자 입니다")
         }
-
+        if (CryptoUtil.comparePassword(password, userResult.password) === false) {
+            throw Error("비밀번호가 틀립니다")
+        }
         return userResult
     },
 
     //비밀번호 초기화
     resetUserPassword: async (connection, email, changedPassword) => {
-        const userResult = await UserDao.resetPasswordByEmail(connection, email, changedPassword)
-        if (!userResult || userResult.length < 1) {
-            throw Error("사용자 정보가 없습니다")
+        const userInform = await UserDao.findUserIdByEmail(connection, email)
+        if (userInform.length == 0) {
+            throw Error("존재하지 않는 사용자입니다.")
         }
-
+        const userResult = await UserDao.resetPasswordByEmail(connection, email, changedPassword)
+        if (userResult.changedRows == 0) {
+            throw Error("변경 실패했습니다")
+        }
+        //K1ZB0IHKHM
         return { result: true }
     },
     //비밀번호 변경
     changeUserPasswordConnection: async (connection, password, changePw, userId) => {
-        const userPassword = await UserDao.findPasswordById(connection, userId)
-        if (CryptoUtil.comparePassword(password, userPassword.password) === false) {
+        const userInform = await UserDao.findPasswordById(connection, userId)
+
+        if (!userInform || userInform.length === 0) {
+            throw Error("존재하지 않는 사용자입니다.")
+        }
+        if (CryptoUtil.comparePassword(password, userInform.password) === false) {
             throw Error("현재 비밀번호가 일치하지 않습니다")
         }
 
@@ -98,7 +115,9 @@ module.exports = {
         const { userId } = request
         //db에서 찾아온 이메일 있어야 함
         const userEmail = await UserDao.selectEmailByUserId(connection, userId)
-
+        if (!userEmail | (userEmail.length == 0)) {
+            throw Error("존재하지 않는 사용자 입니다")
+        }
         await UserDao.deleteAuthCodeByEmail(connection, userEmail)
         await DocumentDao.deleteDocumentByUserId(connection, userId)
         const userResult = await UserDao.deleteUserByid(connection, userId)
